requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.9
Flask-Bootstrap==5.2.3
Flask-Login==0.6.2
Flask-CKEditor==1.5.1
Werkzeug==2.3.4
Flask-WTF==1.1.1
stripe==5.5.0
python-dotenv==1.0.0
